pyproject.toml
[project]
name = "llm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.1.0",
]
